refs task #39197 Register
Add registration feature for employee and employer

Create RegisterController with the following methods:
* showRegistrationForm to return a view with registration forms.
* registerEmployee to handle request to register an employee user.
* registerEmployer to handle request to register an employer user.
Assign routes for these above actions.

Validate the request input using form request.
When received a request with valid input, create a user and a profile
for this user.
Set the status of employee user to 'activated'.
Set the status of employer user to 'inactivated'.

Update localization for login and register pages.